Refactor the metrics reporting initializing.
