+++ src/Exceptionless.Web/Bootstrapper.cs
@@ -1,9 +1,5 @@
 ﻿using System;
 using System.Threading.Tasks;
-using App.Metrics;
-using App.Metrics.AspNetCore;
-using App.Metrics.Formatters.Prometheus;
-using App.Metrics.Formatters;
 using AutoMapper;
 using Exceptionless.Core;
 using Exceptionless.Core.Extensions;
@@ -11,14 +7,12 @@ using Exceptionless.Core.Jobs.WorkItemHandlers;
 using Exceptionless.Core.Models;
 using Exceptionless.Core.Models.Data;
 using Exceptionless.Core.Queues.Models;
-using Exceptionless.Insulation.Metrics;
 using Exceptionless.Web.Hubs;
 using Exceptionless.Web.Models;
 using Exceptionless.Web.Utility;
 using Exceptionless.Web.Utility.Handlers;
 using Foundatio.Jobs;
 using Foundatio.Messaging;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
@@ -26,30 +20,6 @@ using Stripe;
 
 namespace Exceptionless.Web {
     public class Bootstrapper {
-        public static void ConfigureWebHost(IWebHostBuilder builder) {
-            if (!String.IsNullOrEmpty(Settings.Current.ApplicationInsightsKey))
-                builder.UseApplicationInsights(Settings.Current.ApplicationInsightsKey);
-
-            // Note: The prometheus reports metrics in passive mode, so it should only used in webapps but not in console apps.
-            if (Settings.Current.EnableMetricsReporting) {
-                if (Settings.Current.MetricsConnectionString is PrometheusMetricsConnectionString) {
-                    var metrics = AppMetrics.CreateDefaultBuilder()
-                        .OutputMetrics.AsPrometheusPlainText()
-                        .OutputMetrics.AsPrometheusProtobuf()
-                        .Build();
-                    builder.ConfigureMetrics(metrics).UseMetrics(options => {
-                        options.EndpointOptions = endpointsOptions => {
-                            endpointsOptions.MetricsTextEndpointOutputFormatter = metrics.OutputMetricsFormatters.GetType<MetricsPrometheusTextOutputFormatter>();
-                            endpointsOptions.MetricsEndpointOutputFormatter = metrics.OutputMetricsFormatters.GetType<MetricsPrometheusProtobufOutputFormatter>();
-                        };
-                    });
-                }
-                else if (!(Settings.Current.MetricsConnectionString is StatsDMetricsConnectionString)) {
-                    builder.UseMetrics();
-                }
-            }
-        }
-
         public static void RegisterServices(IServiceCollection container, ILoggerFactory loggerFactory) {
             container.AddSingleton<WebSocketConnectionManager>();
             container.AddSingleton<MessageBusBroker>();
