+++ src/Exceptionless.Web/Program.cs
@@ -1,5 +1,9 @@
 ﻿using System;
 using System.IO;
+using App.Metrics;
+using App.Metrics.AspNetCore;
+using App.Metrics.Formatters;
+using App.Metrics.Formatters.Prometheus;
 using Exceptionless.Core;
 using Exceptionless.Core.Utility;
 using Exceptionless.Insulation.Configuration;
@@ -79,11 +83,29 @@ namespace Exceptionless.Web {
 			if (useApplicationInsights)
                 builder.UseApplicationInsights(Settings.Current.ApplicationInsightsKey);
 
-            if (settings.EnableMetricsReporting && settings.MetricsConnectionString is DefaultMetricsConnectionString) {
+            if (settings.EnableMetricsReporting) {
                 settings.MetricsConnectionString = MetricsConnectionString.Parse(settings.MetricsConnectionString.ConnectionString);
-                Bootstrapper.ConfigureWebHost(builder);
+                ConfigureMetricsReporting(builder);
             }
             return builder;
+        }
+
+        private static void ConfigureMetricsReporting(IWebHostBuilder builder) {
+            if (Settings.Current.MetricsConnectionString is PrometheusMetricsConnectionString) {
+                var metrics = AppMetrics.CreateDefaultBuilder()
+                    .OutputMetrics.AsPrometheusPlainText()
+                    .OutputMetrics.AsPrometheusProtobuf()
+                    .Build();
+                builder.ConfigureMetrics(metrics).UseMetrics(options => {
+                    options.EndpointOptions = endpointsOptions => {
+                        endpointsOptions.MetricsTextEndpointOutputFormatter = metrics.OutputMetricsFormatters.GetType<MetricsPrometheusTextOutputFormatter>();
+                        endpointsOptions.MetricsEndpointOutputFormatter = metrics.OutputMetricsFormatters.GetType<MetricsPrometheusProtobufOutputFormatter>();
+                    };
+                });
+            }
+            else if (!(Settings.Current.MetricsConnectionString is StatsDMetricsConnectionString)) {
+                builder.UseMetrics();
+            }
         }
     }
 }
